refactor(user-playlist): add user logged check to create, update, and delete user playlists

+++ src/services/UserPlaylistService/UserPlaylistServiceImpl.ts
@@ -10,6 +10,8 @@ import {
   MusicDTO,
 } from '@/models/playlists';
 import { createLogger } from '@/lib/logger';
+import { AuthContext } from 'contexts/auth-context';
+import { NotFoundError } from '@/models/Errors';
 
 const logger = createLogger('PlaylistService');
 
@@ -19,23 +21,47 @@ export class UserPlaylistServiceImpl implements UserPlaylistService {
   ) {}
 
   public async createPlaylist(
+    userId: string,
     data: CreateUserPlaylistDTO
   ): Promise<UserPlaylistDTO> {
-    logger.debug({ userId: data.userId, name: data.name }, 'Creating playlist');
-    const result = await this.userPlaylistRepository.create(data);
+    logger.debug(
+      { name: data.name, privacity: data.privacity },
+      'Creating playlist'
+    );
+
+    const result = await this.userPlaylistRepository.create(userId, data);
+
     logger.info({ playlistId: result.id }, 'Playlist created');
+
     return result;
   }
 
   public async updatePlaylist(
+    userId: string,
     id: string,
     data: UpdateUserPlaylistDTO
   ): Promise<UserPlaylistDTO | null> {
+    const existingPlaylist =
+      await this.userPlaylistRepository.findByIdAndUserId(id, userId);
+    if (!existingPlaylist) {
+      logger.warn({ playlistId: id }, 'Playlist not found for update');
+      throw new NotFoundError('Playlist não encontrada');
+    }
+
     return await this.userPlaylistRepository.update(id, data);
   }
 
-  public async deletePlaylist(id: string): Promise<boolean> {
+  public async deletePlaylist(userId: string, id: string): Promise<boolean> {
     logger.debug({ playlistId: id }, 'Deleting playlist');
+
+    const existingPlaylist =
+      await this.userPlaylistRepository.findByIdAndUserId(id, userId);
+
+    if (!existingPlaylist) {
+      logger.warn({ playlistId: id }, 'Playlist not found for deletion');
+      throw new NotFoundError('Playlist não encontrada');
+    }
+
     const result = await this.userPlaylistRepository.delete(id);
     if (result) {
       logger.info({ playlistId: id }, 'Playlist deleted');
